r/group_configuration: introduced new configuration update strategy
Introduced a new configuration update strategy that adds an additional
step in order to overcome an issue with availability during
reconfiguration.

The new strategy leverages newly added `cofniguration_update` type
representing nodes to be added and removed from current configuration.
All the nodes that are about to be added to the configuration are first
added as learners. When all nodes are promoted to the voters and if some
of the nodes have to be removed the configuration enters joint state and
all the nodes that are supposed to be removed are finally removed.

For more details see:

https://docs.google.com/document/d/1jFGveXNx1HBzFq1r_8MzcdzYVmcZsYaJ_0_a9ZZQ73A

Signed-off-by: Michal Maslanka <michal@redpanda.com>