c/backend: fixed force aborting configuration change
When there is an active leader we must wait for it to dispatch the
finish update command. Otherwise it may be the case that one of the
followers will dispatch the finish update command while leader may
decide to truncate its log. This would lead to situation in which raft
group configuration is not up to date with metadata.

Signed-off-by: Michal Maslanka <michal@redpanda.com>