cloud_storage/tests: collect and use config batch sizes
when comparing segment sizes which are restored from s3, the restored
size is usually smaller than the original size because of the presence
of config batches in the original which are removed from restored
segment.

with this change we try to estimate the size of these omitted config
batches to account for the difference we will tolerate in original and
restored segments.

this is an approximate comparison, it computes the max offsets for a ntp
and uses it to come up with a diff assuming 1024 bytes size of a config
batch - which is actually smaller, close to 512 bytes.