Enable oncore checking in allocation_units.
Even though they ave movable, allocation_units objects must be
destroyed on the same core the *original* allocation was made on,
as they contain an embedded pointer back to the allocation state
which lives on that core, since to do otherwise would be a cross-shard
race-condition.

Enforce this condition in debug mode using oncore checker.