Optimizations to metadata path
We see out of memory errors on the metadata path for large
partition counts. One problematic place would have 3 or 4 copies of the
partition list in flight at once.

This change avoids this code entirely in the usual case that the
metadata request isn't having a side effect of creating new topics
and reduces copies even if it is.

Issue #5563.