cluster: Recovery with information stored in topic_table
Previously, on every replica of every partition in the topic recovery
process downloaded topic_manifest.json as a first step. This step is
removed now. The topic_manifest.json is downloaded in topics_frontend.
The 'remote_topic_properties' struct used by read replicas is repurposed
to store data needed by the recovery process. The recovery is never used
on read replica topics so it's safe.