tests: lower the max metadata age
This commit sets the maximum metadata age to 5s for the
test_old_node_join test. The test asserts that all brokers have
acknowledged a new cluster member via the admin API. The current timeout
for this assertion is set to 10s and the maximum metadata age was also
set to 10s (by default). The assertion and the metadata refresh could
end up racing which results in a misleading test failure.