cloud_storage/tests: use delta of upto 4k
files on disk can be larger than actual data uploaded to cloud storage.
it appears this is due to block allocation of segments, in this case the
maximum diff can be 4095 bytes, if there is an extra block allocated for
one byte on disk. so we adjust the margin of error to 4096 bytes to
account for these differences.