test/read replica: validate replicas don't write to s3
Checks total number of objects in bucket, and the total bytes of data in
the bucket. Repeats measurements while read replica is running, to
confirm no objects nor bytes are added to the bucket.

I add these checks to the existing end-to-end test to save some time;
these are fairly slow tests to set up.