partition_balancer: fix leader_id if leader not stable
If the node is a provisional controller leader and it hasn't
yet committed the first batch in its leadership term, local raft0 doesn't
consider itself as leader but returns self id from leader_id, so we
return it from partition_balancer_backend. It leads to self rpc request
when we are trying to proxy request to leader.
This commit removes this situation