cluster: improve shutdown logging
For classes that have an asynchronous ::stop
method, it is very helpful to log on entry, so that
in the event of a shutdown hang bug we have a good chance
at telling which shutdown stage had the problem.

Logging at INFO is a little verbose, but it is only
one time on shutdown & if a user hits a shutdown
hang they are unlikely to have debug logs turned
on in advance.