tests: don't ignore shutdown hangs in failure injector
Previously, a hang on SIGINT would be logged, but then
the test would proceed with a node in this zombie state.

Since shutdown hangs are always a bug, it is correct to fail
the test as early as we can with the right reason, so just
re-raise the exception instead of trying to proceed.

Related: https://github.com/redpanda-data/redpanda/issues/5178