redpanda_installer: don't initialize releases unless needed
Previously when the RedpandaInstaller starts up, we would pull a list of
available released versions from Github. This is useful when looking for
valid versions to upgrade from (e.g. when using
`highest_from_prior_feature_version()`), but is not required for every
usage of the installer.

When running upgrade tests hundreds of times to snuff out flakiness, we
may end up needlessly hitting Github's API caps[1]. This commit updates
the behavior to only initializes the releases list when needed.

[1] https://docs.github.com/en/rest/overview/resources-in-the-rest-api#rate-limiting