Remove size restrictions on frag_vector elem type
fragmented_vector is a vector-like class which limits the maximum
contiguous allocation required by breaking the elements up into chunks
of a specific maximum size call the "fragment size".

Currently, the vector places two restrictions on the fragment size:
 - It must be a power of two
 and one on the element size:
 - It must evenly divide the fragment size.

It seems to be that neither of these is necessary, and they severely
limit the element types that may be used: effectively, only elements
whose size is a power of 2 can meet these requirements.

The first restriction may be in place to ensure the fragment size lines
exactly with seastar span sizes, and the second may be so that there is
no internal waste within a fragment, but with many object sizes this
waste is very small.

We simply remove these restrictions for now.