r/tests: fixed incorrect assumptions in replace_whole_group test
Fixed `replace_whole_group` test by waiting for the whole raft group to
be up to date before executing configuration change.

The test was flaky as sometimes the configuration was changed before
propagating any records to one of the nodes (raft correctness is based
on majority). This caused the condition validating if some batches are
read to fail permanently.

Fixes: #342

Signed-off-by: Michal Maslanka <michal@redpanda.com>