admin: include all members in "/brokers" response
This commit changes the handling of the "/brokers" admin request to
include all brokers present in the members table in the response.
"cluster_view" has the same behaviour.

Previously, only brokers returned by the report generated from the
health monitor were included. In that case, to return all brokers the
health metadata had to be fresh. This meant that a new joiner, would not
be present in the "/brokers" response until the node updated it's health
metadata.