tests: add explcit check of config status endpoint
I noticed when I broke something in the status reporting,
that only one of the rpk tests failed, because there wasn't
anywhere else we checked the overall length of the status
vector.

Add an explicit check for this as part of the basic GET check.

(cherry picked from commit 49be9abf536940ed51123af41b8778de25480550)