archival: Always stop the ntp_arhiver on shutdown
If an `ntp_archiver` is added, and the service is stopped prior
to completion of manifest download, the `ntp_archiver` will not be,
stopped, and its gate not waited on.

Fix that by always adding the `ntp_archiver` to `_archivers`.

Fix #4601

Signed-off-by: Ben Pope <ben@redpanda.com>
(cherry picked from commit f10c42ca21c3d4c8ba618b16f2373a322186879f)