cluster: enable controller replay if last_applied is ahead of log
This situation happens if someone deletes controller log
segments while leaving the kvstore in place.

Previously, the kvstore last_applied would cause the
node to hang waiting for the controller log to replay
to that offset.

Now, we log an error about the apparent inconsistency,
and proceed.

In general we do not want to ignore data inconsistency, but
this is a special case: deleting the controller log is something
a user might legitimately do in order to work around another
issue + force redpanda to rebuild the local copy of the
controller log.