controller: use app-wide abort source in ::start
This fixes shutting down redpanda while it is busy
replaying controller log.

The controller has its own abort source, but the
it does not get registered as a defer() hook (via shutdown_input)
until after start() has returned.

To enable long running parts of start() to be interruptible,
we need an external abort source.  Conveniently application
already has one: it is only on shard 0, but that's okay because
we only need it on controller_stm_shard (0).