cloud_storage/tests: wait for partition metadata
If a leader election happened right before we verify topic watermark,
rpk may return an error such as "not leader for partition" due to stale
metadata.

Instead of failing immediately this change makes the caller wait for
upto 1 minute so that the new leadership info is propagated to the
cluster and rpk returns the correct partition metadata.