cloud_storage: use is_elected_leader for leader check
during segment upload we check for leadership lost. this change moves
the query method from is_leader to is_elected_leader. is_elected_leader
has weaker requirements and is used in the rest of ntp archival. as the
rest of the system moves to is_leader, this check will also be moved. as
of now a mismatch in the queries causes uploads to cancel.