schema_registry: Refactor due to kc::list_offsets change
`kafka::client::list_offsets` now throws an exception if the
response doesn't contain the expected single topic_partition,
or if the tp has an error, and retries didn't mitigate the
error.

That exception type is suitably handled by Schema Registry,
as it's a base of kafka::exception.

Remove explicit error handling checking and handling from
call-sites in Schema Registry.

Signed-off-by: Ben Pope <ben@redpanda.com>