cluster: Call archival_metadata_stm sync in upload loop
The archival STM can lag behind too much because it calls 'sync' method
only after segment uploads. This can cause timeouts. To prevent this
problem this commit makes 'sync' method public in the
'archival_metadata_stm' and calls it periodically from the upload loop
even if there is nothing to upload yet. This will guarantee that the STM
is not lagging behind too much.