Add optional fragmentation resistance to schemata
Certain responses may be very large, such as a metadata response on
a cluster with many partitions. Since int schemata we map lists in
response types to std::vector, this may result in large contiguous
allocations, which is a no-no in redpanda: we should limit such
allocations to 128 K.

To fix this, we add an enable_fragmentation_resistance list in the
generator: array of types in this list will be implemented with a
large_fragment_vector instead of std::vector: this type limits the
contiguous memory allocated by making several smaller allocations if
necessary.