Log node status cache update contents
There is a class of test failures where a follower's health monitor is
apparently unaware of one of the nodes in the cluster. The existing
logging is insufficient to trace whether the node status cache has been
updated or not. This commit adds the log line that makes it possible.