scale_tests: added partition balancer scale test
Added a test that test partition balancer operation with large number of
partitions in cluster.

The test creates a 5 node cluster with 18000 partitions in it. Then
makes one of the nodes unavailable to trigger partition rebalancing. The
test is finished when there are no replicas on the node which was
stopped.

Signed-off-by: Michal Maslanka <michal@redpanda.com>