test: add alternate wait util for segment removal
This commit introduces a new test utility for waiting for the removal
of a partition's segments, wait_for_removal_of_n_segments. It
periodically request a snapshot of the segments associated with a given
partition and compares it with the provided original snapshot.

As opposed to wait_for_removal_of_segments, its result is not impacted
by newly created segments. This means that it can be safely used in
contexts that produce simulated failures (each failure causes the
current segment to roll).

(cherry picked from commit 5601891389afc7c074e437f7399a48d3f28079d1)