tests/si: use snapshots to find removed segments
Previously, the shadow indexing end-to-end asserted against the current
number of segments when checking for segment removal. This approach has
the downside that a restart/failure of a redpanda node causes a segment
roll, which makes the assertion unreliable in a context with simulated
failures.

This commit changes the assertion to use the
wait_for_removal_of_n_segments helper method which uses segment
snapshots to determine how many segments were removed. The change
deflakes the test against a high number of injected node failures.