cluster: Fix snapshot initialization for topic recovery
After topic recovery redpanda initializes raft snapshot and then it
starts raft bootstrap. To create the snapshot we need to provide
last_included_index field. Consensus can truncate configuration up to
last_included_index (and also truncate offset translator state as well)
so when we creating the snapshot we're setting this value to
model::prev_offset(min-recovered-offset) to be safe.

The problem is that group_configuration with offset model::offset{} is
always added in c-tor of the configuration_manager. Also, the
model::prev_offset returns model::offset{} for offset 0. So if the log
downloaded starting from offset 0 we will create a snapshot with
last_indcluded_index set to model::offset{} which could lead to
conflict (we can't add this config to configuration_manager because the
offset is already used).

The fix changes model::prev_offset to get_prev_offset function which
returns -1 for 0 value. In this case there is no conflict.