c/partition_balancer: introduce hard_max_disk_usage_ratio
When moving partitions away from unavailable nodes it is desireable to
violate disk ratio that we use to determine when we need to move nodes
away from a node - this will allow us to save data if we have some free
space left. But we shouldn't go to 100%, so we use the
storage_space_alert_free_threshold_percent config value to determine the
hard limit.