upgrade_test: reduce workload throughput
Tests that run workloads through a rolling restart are susceptible to
flakiness when the node being restarted is knowingly slower than the
others. This can typically the case when running in mixed versions,
wherein one node is the locally-built debug binaries while the others
are downloaded release binaries. Upon returning from maintenance mode,
we attempt to wait for the number of recovering replicas to drop to
zero, but even the reduced workload we're running leads the test to be
flaky ~5% of the time when run locally.

This commit significantly reduces the workload sent, while still
ensuring progress is made through the rolling restart. I ran
UpgradeWithWorkloadTest.test_rolling_upgrade_with_rollback 100 times
with this reduced workload and saw no failures.

Fixes #5713