cluster: Do not generate raft snapshot if recovered empty log
Since raft snapshot for recovered partitions is added at offste -1 we
can trigger an assertion if the downloaded offset range is empty.
In this case the dirty offset of the log will be model::offset{} but the
last_included_index will be set to -1. This violates one of the raft
invariants (snapshot can't have larger index then the dirty offset of
the log). To avoid this problem this commit adds extra check for this
case. If downloaded log doesn't have any data the code doesn't create a
snapshot (or any other state) and proceeds. In thi case it just creates
an empty partition.