tests: wait for node to be reported alive in rack awareness test
Rack awareness is forced in redpanda with soft allocation constraint. It
is done in such a way to make it possible to allocate partitions even if
there are not enough nodes in distinct racks. When the rack awareness
test was running the partition balancer was able to calculate movements
before the node that was previously stopped was reported alive. This way
a partition must have been allocated when two nodes were unavailable
when those two nodes happened to be on the same rack the rack awareness
constraint could not be held.

Fixes: #5795

Signed-off-by: Michal Maslanka <michal@redpanda.com>