cloud_storage: Handle empty segments in recovery
Configuration record batches are removed from the downloaded segments
during topic recovery. Because of that downloaded segment could become
empty. It's also possible that the entire log will become empty for the
same reason. In this case we need a way to detect this and to reset
start offset of the log.