rpk group: fix describe
Commit 15df1b34fac1b63b14d0d85eefd1bfd410e764bb broke describe.
Previously, if a group had no active members, describe would list
offsets for *all* topics and partitions in the cluster. This was
wasteful and would be slow on huge clusters. CalculateGroupLag would
then filter for only topics that have commits, and then print only what
was necessary.

Describe was changed to only list offsets on groups that were consuming.
Unfortunately, this broke empty groups.

We now describe a union of all topics that have commits, and all topics
that a group is consuming. This is the minimal amount of topics needed
to satisfy the group lag calculation and does not waste extra resources.

We bump kadm to take advantage of a new API to keep this code nice.