compat: make negative check tests quiet
Prior to this change compat tests printed diagnostics to standard out.
With the introduction of the negative compat checks all the tests are
expected to fail in some way, which leads to a lot of noise. This commit
packages up diagnostics into an exception which lets us easily control
how that output is dumped.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>