tests/services: workaround kgo-repeater consumers disappearing sometimes
I think this is a bug with the workload generator (or, unlikely perhaps
a problem with franz-go).  It is usually only a few consumers that disappear
from the group, so it doesn't hurt the validity of the overall scale
test, and we can hunt it down separately.

(cherry picked from commit c17e2cb6320ea34317a79c8f98de93ab1996287a)