test: wait for cluster config status on controller leader
If we query another node, it is possible to see persistent
status updates for the nodes _other_ than the one we are
querying, and non-persistent update to the stauts of the node
we are querying, that passes the version check.

Then if we query status on a different node a moment later,
we will see an older state for the node we first queried.

This only matters for tests that are actively trying to read the
status _again_ after wait_for_version_sync.  wait_for_version_sync
was already correct inasmuchas when it complete the config has
been applied everywhere.

(cherry picked from commit 13d211adc4a0cbe2d4b6126d7f21717e85ecc877)