cloud_storage: use larger cache for scale test
The cloud storage cache used to be cleared in regular intervals. This
allowed the test to actually grow the cache to larger than limit in
short bursts between cleanup, and pass because the consumers could
progress.

With the new more strict/realtime cache eviction, this does not happen
and the test fails for multiple reasons. This change allows the cache
size to be a multiple of segment size so that there is not a continuous
cycle of hydrate current segment -> read from segment -> evict old
segment in the code.