rpk: sandbox mode: default and container mode
This commit introduce 2 options for sandbox mode:

Defailt: Intended for laptop use.
Container: ...

--sandbox is an alias to --sandbox=default

We are now in the path of deprecating
redpanda.developer_mode, so if its enabled
we will default to --sandbox=default.

Sandbox localhost include overprovisioned, thats
why we are also adding overprovisioned to the
config.Default() method.