c/health_monitor_backend: use consistent refresh strategy on all nodes
Changed the way how cluster health is refreshed on the `raft0` leader
node. Currently the refresh will be triggered if controller leader
health metadata is stale. Previously the refresh was timer base it might
have happened that when leadership was change requester accessed a stale
health metadata which were much older than `max_metadata_age`.

Now every time the health report is requested we check the metadata age
and if metadata is stale we dispatch an appropriate refresh request
either contacting leader or gathering information from the cluster.

Signed-off-by: Michal Maslanka <michal@redpanda.com>