txn coordinator: fix split-brain violation
Kafka API doesn't have explicit begin txn API, when a transaction
coordinator recieves first add_partition or add_group it starts a
transaction. Also Redpanda defers disk flushes to the commit moment.
Combinations of those thing caues a problem:

  1. a client issues add_partition to txn coordinator
  2. txn coordinator starts a transaction (this is in memory state)
  3. the client writes a message to data partition
  4. redpanda treats the write as acks=1 and acks the request before
     the replication is finished (it's safe to do it because on commit
     redpanda checks that all pending replication is done)
  5. txn coordinator & data partition experience re-election and the
     in memory state is lost
  6. the client issues add_group to txn coordinator
  7. txn coordinator is unaware about the ongoing txn and starts new
     transaction
  8. the client commits the txn and only the consumer group change is
     written

Since the data record was written with acks=1 just before re-election
it gets lost and the client hasn't figured out that there was some-
thing wrong with transaction.

The fix is to write to the txn coordinator's log when a transaction
starts; in this case when a crush-induced re-election happens new txn
coordinator has an opportunity to detect an ongoing txn and fail it.