cloud_storage: adjust condition to remove archiver
an archiver is removed in reconcile loop if it is stopped. the stop
condition checks if the upload loop is stopped. with read replica there
are two status variables which denote a stopped loop. this change makes
the condition such that any of these two variables being true marks the
archiver as stopped, as the two conditions start out being false. they
turn to true in mutually exclusive code paths.

the archiver must be marked as removed in order for a stopped upload
loop to restart.

(cherry picked from commit 5d93dda84394dea4c77893e88eb40459ecd9c746)