tests/rptest: disable iam role tests for ec2
IAM role tests use a bundled mock server script which has a fixed set of
credentials which match our MinIO docker image.

These tests do not work with real S3 because it expects credentials
which are supposed to be part of the test context. The tests can be made
to work with S3 by passing the credentials from test context to the mock
server script, but these should be disabled until that change is made.