tests: allow exceptions when checking for health
In waiting for a cluster to become healthy, it's possible that some
nodes haven't started or are in the process of starting up. In such
cases, it's reasonable to expect the check for metrics fails. This patch
updates the health check in RedpandaService to allow such cases.

(cherry picked from commit 34b3bd24ddec207224df880da43e36f46355865b)