cluster: fix spinning in feature manager
The _update_wait wait is conditional on !_updates.empty(). If
there is something in _updates we don't wait. Then in
do_maybe_update_active_version, we don't clear _updates
if we aren't the leader.

The only symptom is that we busy spin + if you have debug
logging switched on, you'll see a spew of "Checking for
active version update..." messages.

Fixes https://github.com/redpanda-data/redpanda/issues/5982