tests/Redpanda: add search_log_all
This commit adds a helper method, named search_log_all(), that searches
for a string on all Redpanda logs. The method returns True when a match
on all logs is found. This follows a similar design to python's
built-in all() function.

This commit also renames the previous search_log() method to
search_log_any() since it follows a design similar to python's built-in
any() function.