s/segment: introduced segment generation counter
A `segment::generation_id` is a counter that is incremented every time a
segment is modified (flushed, written to, truncated). When segment is
being compacted we must know if it was truncated as the copy of segment
which is compacted was not affected by mutation.

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit f5d35f0e6a3b41572e5091ce44a90a54a500482c)