s/log: using generation to drop compaction work interleaved with truncation
Using `segment::generation_id` to detect situations in which staging
compacted segment and compacting it interleaved with mutation operation.

When segment is mutated when being compacted we need to drop the
compaction result not to override the mutated segment.

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit 0655a76ec553851ccdf3b18185444c9f7779cd95)