s/offset_to_filepos: return previous batch if request offset is a gap
When truncating segment in the middle we use
`offset_to_filepos_consumer` to establish the physical position in a
file that reflects the offset. The physical position is then used to
truncate the file on disk.

Previously we required for the requested offset to be equal to the batch
base_offset and only then returned position in the file. When segment is
a compacted segment the requested offset may not exists as it might have
been deleted during compaction process. In this situation we should
return last offset and end position of a batch preceding the gap.

Signed-off-by: Michal Maslanka <michal@redpanda.com>