r/consensus: adapted raft truncation error check to storage semantics
As we introduced the new semantics for truncating when truncation
offset lands in the middle of a gap craved in the offset space by
compaction, we need to adapt raft truncation result check.

Now we log not expected truncation result with warn severity and check
if an update of flushed offset is needed.

Signed-off-by: Michal Maslanka <michal@redpanda.com>