cloud_storage: Save tx-manifest to SI cache
Currently, tx-manifests are downloaded every time. Because of that we
will make a request to S3 when the remote_segment is materialized.
Even if transactions are not used we will still have to make a request
to get NoSuchKey response. To avoid this this commit adds logic that
seralizes the manifest to SI cache.