tests: speed up ClusterConfigTest
The initial wait for cluster_id to be set was not
necessary for most of the tests: instead, just disable
the auto-setting of cluster_id by setting it to a placeholder
value in the initial configuration.

This happens to also fix an issue that occurs when
wait_until is improved, and the previous wait for
config version 2 was not long enough: the wait no
longer exists.

Fixes: https://github.com/redpanda-data/redpanda/issues/6008
(cherry picked from commit 442ac911504438db9528b20f5c17432213bd41d6)