logging: Add a log on completion of shutdown
Sometimes shutdown can be hung and it is not obvious whether
it has finished or not when dealing with a large number of raft groups.
Having a log at shutdown makes it easy to analyze the timestamps.

(cherry picked from commit 1e3473a7d7e2333cb37bb32236b91f884d0b8353)