tests/services: fix KafkaCliConsumer.wait()
This subtly regressed in 20d7ba5cf, which changed
stop_node to use a SIGTERM instead of a SIGKILL.

Because _worker drops out as soon as _stopping is seen,
the remote process may still be running when the
background thread completes, therefore may still
be running after wait() completes.

This allows consumers to run beyond the end of the test
where they were running, and interfere with the start
of the subsequent test.