tests: fix ClusterConfigTest waits
The wait_for_version_sync was correct for
waiting for the configuration to propagate
across the cluster, but it was _not_ correct
for waiting for the configuration status to
be symmetric on all nodes (i.e. for all nodes
to know the status of all other nodes)

For test cases that query the status via
arbitrary notes, a stricter wait is needed.

This issue becomes visible once wait_until is
improved to avoid spurious extra sleeps, in
https://github.com/redpanda-data/redpanda/pull/6003

Fixes https://github.com/redpanda-data/redpanda/issues/6010