tests: speed up acls test setup
It is not necessary to wait for the mtls feature,
as it is no longer used to gate enabling the functionality
in 22.2.

The sleep(5) hack for waiting for ACLs is replaced
by writing a phony user after the acls, and then waiting
for that user to be visible everywhere, as evidence that
the controller log has proceeded past the ACLs.

(cherry picked from commit 5369699946c82a385299caa6d1f5807a66e1f9e0)