tests/services: avoid reading back config files in brokers()
This was redundant: we write out the config file and
then SSH out to the node to read it back each time
we build the brokers list for running a client.

This is a simplification and should also improve
execution time slightly.

(cherry picked from commit 1b75fd60ceac1c22f9e40803a4f20290019f72a3)