http/client: overrides fmt formatter for request
An http request can contain headers which have sensitive information
such as authorization. We redact these fields right now using a special
function which creates a copy of the request with sensitive fields
removed. This has the downside of making the developer remember to call
the function each time the request is printed.

The changes here override the fmtlib formatter for request so that any
objects logged via fmt are automatically redacted. It is not possible to
override the ostream operator without subclassing the request class and
using that subclass everywhere, using the fmt API is less intrusive to
the code base.