s/offset_to_filepos_consumer: fix translating gap offsets
When start offset and truncation offset both were in the gap the
`storage::offset_to_filepos_consumer` was unable to translate the offset
to physical position in a file causing the test to fail.

Changed implementation to let the consumer decide on when to stop
iterating over the set of batches. Previously the stop point was based
on the reader configuration. Now the consumer validates target offset
against the batch offsets and decide whether it should stop iteration
and return physical position. Current implementation retained validation
of requested offset i.e. it is only allowed to request physical position
of batch base offset or offset must be in a gap.

Fixes: #6063

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit dcd6bd96b290e134173d7db6a21d1f575fb7d30b)