kafka: write store_offset before group_commit_tx
It is fix for https://github.com/redpanda-data/redpanda/issues/5163.
Problem is group_*_tx contains only producer_id in key, so compaction
save only last records for this events. We need only save in logs
consumed offset after commit_txs. For this we can write store_offset
event together with group_commit_tx.

Problem: this 2 events contains different record batch type. So we can
not put it in one batch and write on disk atomically. But it is not a
problem, because client got ok for commit_request (see
tx_gateway_frontend). So redpanda will eventually finish commit and
complete write for both this events.