s/disk_log_log: retry truncation when segment generation changed
Truncation of a segment is a two step process in a first step we use
`offset_to_filepos_consumer` to determine a position in a file where the
truncation should happen. The consumer returns a physical offset in a
file. After that the file is truncated. Since both of those operations
are executed as asynchorous actions it may happen that the segment will
be swapped in between the two steps (as a result of adjacent segment
compaction). Added checking of segment generation before the segment
file is truncated to make sure that is wasn't swapped with other
segment. If the swap took place we retry truncation with the same base
offset it will determine a new physical position in a file and truncate
it.

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit 11efead9a799000c241227200d376c06b7523554)