tests: make redpanda methods more thread-safe
Several methods that run for a particular node check that the node is
started. This is somewhat reasonable as we typically expect to want to
run get info for just the started nodes. However, this doesn't play well
if calling any of these concurrently with a restart (removes from the
list of started nodes, then re-adds).

This commit makes the methods more thread-safe by checking the node is
in the immutable full set of nodes, rather than just the started nodes.
We do lose out on preventing some potential programmatic errors here,
but it seems worth the slightly improved thread safety.

Fixes #5827