tests: back-to-back upgrade test
This commit adds a test that performs two upgrades while running some
workloads concurrently.

It's worth noting that, as of writing, two versions ago (v21.11) we
didn't support maintenance mode (or more broadly, rolling restarts), so
if the initial version is on this version, the test will stop the
produce workload before performing upgrades.

I've also parameterized the test to also exercise a single version
upgrade, optionally replacing the first upgrade with a simple rolling
restart.