cloud_storage/tests: filter out segments missing in manifest
When comparing size of restored partition with size on s3, we adjust the
original partition size by deducting non data batches, because these are
not restored to disk.

When computing the size of non data batches we iterate over all segments
and add up all non data batches in them. This sometimes results in
adding up batches from segments which were not part of the partition
manifest and therefore are not going to be restored.

The end result is that the baseline gets deducted by an inaccurate (too
large) amount and the assertion fails indicating that the restored
partition is larger than original.

This change introduces an S3 view which allows to query if a segment is
part of manifest, by matching the path and the archiver term. Only
segments which qualify are scanned for non data batches.

The view can be extended to provide more query functions as needed.